+++ components/common/Button/ExcelButton/styles.css.ts
@@ -1,0 +1,8 @@
+import { style } from '@vanilla-extract/css'
+
+export const button = style({
+    display: 'flex',
+    width: '100px',
+    justifyContent: 'center',
+    gap: '6px',
+})
